[refactor]연령별 차트 컴포넌트 분리
컴포넌트를 따로 작성할 수 있는 부분들은 컴포넌트화 시켜서 코드를 좀 더 확인하기 쉽도록 하였습니다
목데이터의 경우에는 같은페이지의 다른 차트에서 사용될 수 있어서 dashboard로 옮기고,
감염자인원수가 많은 관계로 기준선 2400,2800 을 추가하였고 해당 400명 간격이 50픽셀이기 때문에
한명당 0.125픽셀 즉 감염자수를 8로 나누면 해당연령 stack의 width값을 구할 수 있습니다
기준선과 연령별 stack의 색깔을 나타내는 부분은 수정하기 쉽도록 목데이터를 활용하여 구현하였습니다
